Docs: replace image links in README.md (#351)
What?
This PR replaces broken images link in the README. 

Special notes:
I went to wayback machine to get the original images and saved and
uploaded them again. I could not find in the commit history when/how
they were deleted and didn't think it was necessary to re-create the
original path if they can be hosted this way anyway.